[bashrc] Add a note for vim config init, you need a symlink
Had an issue with syntax highlighting failing to load, and went through
the issue and resolved it. Did not understand that $VIM and $VIMRUNTIME
are so intertwined which means that trying to move our configs breaks
things. While looking to see if I can disable ~/.gnome2/Vim creation,
found that some menus were not loading in gvim, and that was down to
$VIMRUNTIME/optmenu.vim failing to be located.