[lynxrc] Suggest to lynx we prefer UTF-8
Forgot to put lynx on the laptop for holiday, wanted to test something I
worked on then, and got to thinking why it was munging → to be ->, ↑ to
be ^-. Charsets, right?

So opened the best test case,
http://www.cl.cam.ac.uk/~mgk25/ucs/examples/UTF-8-demo.txt but it was
not showing by default in UTF-8, locale is correct, and afaik, terminal
is set correctly for UTF-8. Lynx wrote out ~/.lynxrc of a different
format(!) to the current lynx.cfg I use, copying those values into it
failed to change things. However, changing = to : made it work. Why does
this programme have two styles of config setting?!
Thought a bit more about cookie file, and that moves to XDG_DATA with
a new redundant bookmark file.