[vimrc] Toggle cursorline on inactive windows, enable colourcolumn
Been using (cc) for a week and really liking it to help make me remember
line wrapping. It also looks good on the desktop with the current colour
scheme, but I know some of the colours are a bit jarring – that is
something to look at in the future! But I hate that kind of colour
choice.