[sshconfig] Correct placement for IdentitiesOnly
Turns out the existing settings in the other Host settings is not enough
to make this work - the IdentityFile command looks to just add to the
list, even though we do not hit that with the match. Oh well, enable
this on Host * as well.

What prompted this was https://blog.filippo.io/ssh-whoami-filippo-io/
checking that what I believed I had set correctly was working. It did
not find my github account, as the key is newer than their collected db,
but running ssh -v did highlight that two keys were sent, rather than
just my %l_rsa.