[bashrc] Correctly set BZR variables
Apparently I am an idiot and had these all wrong. It probably does not
help that I have not touched `bzr` since the pull request I made for
Zim, and now with my current ssh config I cannot even bzr if I wanted
to.

Anyhow, we don’t need to set BZRPATH as that is for binaries, we should
not set BZR_HOME because current behaviour is to fallback onto
XDG_CONFIG_HOME with one exception, XDG_CONFIG_HOME/bazaar must exist or
it will use ~/.bazaar still
BZR_LOG needs to be set as bzr ignores XDG stuff if BZR_HOME is not set
and will generate ~/.bzr.log because… no good reason.
And also we don’t need to set BZR_PLUGIN_PATH as we have no plugins.
Revisit in future if a plugin can fix the mess that BZR_SSH doesn’t help
us with.