[bashrc] for now, comment out the alias for `lsof`
Sigh, applications, stop using aliasable commands. Something is
triggering lsof to warn on a faulty -e when we run the tool script. I
have an assumption because it is running a mini shell that bashrc was
being unparsed, but I don’t know /what/ is calling lsof to trigger the
complaint.