[bashrc] Add current directory to our shell window title
This uses an ANSI escape of OSC 0, terminated with BEL, to put our $PWD
(with a s/$HOME/~/ tweak) into the terminal window name

Currently looking if tmux exposes an envvar that would allow
PROMPT_COMMAND to update, but I doubt it would get processed? Seems we
need to look for tmux support for sending output the shell would know.