[tmuxconf] New file, tweaks to tmux
Probably well beyond the point of needing this in our dotfiles repo, it
has had numerous tweaks over the years - whether correct or not.

Removed some chunks from this file that I want to revisit (theming,
mostly), and snipped comments and toggled settings that no longer apply.

Main thrust of this work was because I was trying a new server out, and
noticed tmux had had a few good tweaks since my last use to better
configure it at runtime (interactive status, and command history)

A new thing for me in this is including a timezone in the timestamp, tz
offset I think is an important bit of information however… it’s kinda
hard not to want both bits of info? I know +0100 is great, but that
doesn’t inform whether it is a DST offset or not. BST does, yet doesn’t
give the offset for people unfamiliar with British Summmer Time.

Which makes it feel like +0100(BST) is needed, or another token to
indicate DST.

Blergh.