[bashrc] Add more colour to our PS1
Tooling around in a newer Ubuntu has made me appreciate the blue styling
for the path, and wishing for a slightly more prominent user@host, this
should do it.

It would be nice to be an all-black : exist between user@host and pwd,
to allow for really simple rsync copy/paste operations, however the
usefulness of grabbing a path from shell is more common, so we settle
for a space gap.