[motd] Commit our motd script (Debian‐flavour)
A simple /etc/motd greeter I made in 2011, and then never touched again
until I discovered it was broken on Ubuntu and … never really had the
heart to delve into why. Now that is resolved, though now rather than
being a cat‐able file, it is a Bash script.

Wanted it to show a colourised full domain for the host, kinda centred
on the full stop between shortname and the rest but not really. It has a
bias towards that, but nothing that would make it too lop‐sided.

On top of that, the previous line I typically reserve for a link to my
website, but that didn’t feel generic enough so made that align too

This relies on ansi2txt for stripping $MESSAGE escape codes, which
currently is in `colorized-logs` package but on this machine it was
`ktbin`

Assuming there are portability issues in this script, so expect further
patches. \e \033 ^[, escape codes are quite a mixed bag, huh?