[initialise] Commit first stab at a bootstrap script
This is for ‘new’ systems, to wrangle ~ into somewhat of a comfy,
familiar place.

It creates paths in ~ with `mkdir -p`, to hopefully prevent warnings and
overwrites.
Where coping files, it uses `cp -n` to avoid clobbering existing files.
It uses `sudo` once, to link a newly‐created custom motd script into a
system path.

Plenty of init missing, but want to revisit this again in the future and
it is always better to have bulletpoints along the journey.