[.gitignore] Hide newly created hostname‐specific files from git
There are a few files I don’t commit to this repo for obv. reasons -
password files, ssh pub & priv keys, and the odd machine-local file.

Some are checked into local branches to still maintain state for
changes, even if they are not actually commit to master.

So this is the start of revealing some of that behind the scenes action.