[bashrc] Add PATH references for Rust
Had forgotten to include my local Rust PATH changes… mostly because I
rewrote this to avoid duplication, but then held it back because I was…
unsure this was a good approach.

Will convert these in the future to a non-exported function call to
make it more readable and prevent mistakes.