[initialise] Add git configuration
As explained in 40fc1b6558ccd330f195177712409b1eea3ed938, we copy gitk
into our ${XDG_CONFIG_HOME}/git location, and symlink the remaining
config and ignore files.

We also symlink some bin scripts that are included in this repo.

If in the future we modify gitk (unlikely), will need to recopy over.