[initialise] Correctly set up vim for the current version
vim wants access to the scripts it ships with, however I don’t know a
better way to refer to these things programmatically in vim.

As done here, using globbing to match the filename and them create a
reference to it in our vim config dir. For the second instance, makes a
vimcurrent pointer so our vimrc can refer to it. This might be a lot of
redundant work, if so, my bad.

Also tidy up the runtime parameter in vimrc, seems we had some
duplicates and redundant (symlinked to same source) paths.