[bash_logout] Commit logout script for xauth
To tidy‐up the ~/.Xauthority file, we need to relocate it and set an
envvar, however on next log in, xauth will error with:

/usr/bin/xauth:  file /home/john/.Xauthority does not exist

This is because even though it is a runtime file, and will get rebuilt
if missing, the error is apparently useful.

To (kinda) resolve this, on logout we touch this file, all to prevent
the warning. ~/.bash_logout wont be called for some instances of
disconnecting (like the server shutting down while logged in), so… just
accept it.

This adds changes to initialise to backup the default logout script and
symlink our new one. There is a further action to be taken, and that is
moving the existing ~/.Xauthority script yet we aren’t 100% complete, so
we don’t as-yet source bashrc at the end of initialise. Once that is
done, that action can be done early, or just left for our bashrc